#22-Implement pagination for posting page
It looks more simple than forum page since it's only have previous and next buttons. However, when it's on first posting and last posting button should be disabled and also need to make fake link so the thymeleaf grammer got bit complicated.

To determine the last posting, need to know the total count, so added the coun tquery and add service logic.