#38-<form> Add encoder (for data transfer testing)
Form data must be serialized and used in a different form from the API.
Converts an object into a string in form data format according to the regulations.
Make the utility dedicated to testing, and write tests as well
If you do not have this, you can enter the value in the form of `post().param()`
The expression does not seem to match the actual post request transmission,
It is inconvenient because it must be entered in units of parameters rather than objects.