#62- Apply getter to entitiy comparison logic (equals(), hashcode())
If you access the field directly here,
When dealing with proxy objects created by Hibernate for lazy loading
Field value may be `null`
Then, the comparison logic cannot be performed properly.
Using getter.
With this, you can read the value properly even if you use a proxy object.

Also found that
if (!(o instanceof  Posting posting))  -> didn't change the variable posting to that so fixed it.