#60-Reflection of domain changes: Hashtags
A hashtag entity is added, and a posting page contains multiple hashtags.
Because of this, the changes were widely propagated.
It is not a good idea to leave a commit in a compilation error state.
I have no choice but to put all the changes in one commit.
Let's think more about whether this method will work.
However, because it only reflected domain changes and did not implement any functions,
Only temporarily avoids compilation errors, but the test currently fails.