#72 -Implement hashtag service
Implement a function to manage hashtags on posts.
Uniquely, there is no method that handles the storage of hashtags.
Storage of hashtags is dependent on the structure of the current business logic.
This function only occurs when a post is saved or modified.
This is because everything is processed by the post service through correlation mapping.

* Search the database for hashtag information using the hashtag name.
* Parse hashtags from content
* Delete hashtags that no longer have posts
* TODO: A test for the delete function was not created yet. It would be nice to add it.