#72-Reflection of changes to hashtag function in post service
Mainly when posts are saved, edited, or deleted, describes how hashtags should move together.

Checked how the change of domain impacts to wide range of codes, so to resolve compilation errors  I reflected all the codes including the change of test.

While going through this process there is not enough time to clearly distinguish and show the code writing process.
Thus this kind of commits occured.
In practice, if you divide the work details like this, It may be somewhat inconvenient when tracking history through commits.

Let's ensure that service changes are included with test changes.

* TODO: The logic of `getHashtags()` seems to be able to be improved.