#75- Design domain for child comment
Added code to set parent and child relationships within the comment domian.

Apply all cscading rules so that changes in the collection of child comments are reflected in the query.

This time, we decided to use the one-way association setting. Therefore, the parent comment directly represents the long id rather than the entity.

Additionally, an additional method is provided to add child comments.