#103-Resolving an error that occurs in the main Spring Boot test.
DB access information is replaced with environment variable syntax,
but `@SpringBootTest` reads it as is,
and since it is not in the correct jdbc url format, it causes a failure.

So, I assigned the h2 path, which is an in-memory test db, to the newly created test-only `test` profile, and when running the test, it looked at this profile to solve the problem.